+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Replaced `lazy_static` with the new `std::sync::OnceLock`.
+
 ## [0.12.3] - 2023-07-10
 
 ### Changed
@@ -225,6 +231,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release.
 
+[Unreleased]: https://github.com/aicers/giganto/compare/0.12.3...main
 [0.12.3]: https://github.com/aicers/giganto/compare/0.12.2...0.12.3
 [0.12.2]: https://github.com/aicers/giganto/compare/0.12.1...0.12.2
 [0.12.1]: https://github.com/aicers/giganto/compare/0.12.0...0.12.1